+++ packages/components/src/components/multi-select/multi-select-option/multi-select-option-styles.spec.ts
@@ -2,10 +2,16 @@ import { getComponentCss } from './multi-select-option-styles';
 import { validateCssAndMatchSnapshot } from '../../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
-  it.each<Parameters<typeof getComponentCss>>([['light'], ['dark']])(
-    'should return correct css for theme: %s',
-    (...args) => {
-      validateCssAndMatchSnapshot(getComponentCss(...args));
-    }
-  );
+  it.each<Parameters<typeof getComponentCss>>([
+    ['light', false, false],
+    ['dark', false, false],
+    ['light', true, false],
+    ['dark', true, false],
+    ['light', false, true],
+    ['dark', false, true],
+    ['light', true, true],
+    ['dark', true, true],
+  ])('should return correct css for theme: %s, isDisabled: %s and selected: %s', (...args) => {
+    validateCssAndMatchSnapshot(getComponentCss(...args));
+  });
 });