+++ packages/components/CHANGELOG.md
@@ -14,6 +14,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+#### Fixed
+
+- `Multi Select`: refactor checkbox and fix A11Y test
+  ([#3819](https://github.com/porsche-design-system/porsche-design-system/pull/3819))
+
 ### [3.28.0-rc.1] - 2025-04-14
 
 #### Fixed